+++ helpdesk.tickets-service/src/main/java/innowise/microservice/helpdesk/ticketsservice/exception/FeedbackNotFoundException.java
@@ -4,7 +4,7 @@ public class FeedbackNotFoundException extends EntityNotFoundException {
 
     private static final String DEFAULT_MESSAGE = "Feedback doesn't exist! feedbackId:%s";
 
-    public FeedbackNotFoundException(int feedbackId) {
+    public FeedbackNotFoundException(Long feedbackId) {
         super(String.format(DEFAULT_MESSAGE, feedbackId));
     }
 }
